DA#122: Modified: Rename webViewProvider.ts->metaData.webview.ts and add connectionScreen.webview.ts to get connection screen
