+++ src/extension.ts
@@ -23,7 +23,7 @@ import { IConnection } from "./model/IConnection";
 import { INode } from "./model/INode";
 import { PagerNode } from "./model/PagerNode";
 import { ScopeNode } from "./model/ScopeNode";
-import { getBucketMetaData, getDocumentMetaData } from "../src/webViews/webViewProvider";
+import { getBucketMetaData, getDocumentMetaData } from "./webViews/metaData.webview";
 import ClusterConnectionTreeProvider from "./tree/ClusterConnectionTreeProvider";
 import {
   addConnection,
