+++ src/webViews/metaData.webview.ts
@@ -146,43 +146,3 @@ export function getDocumentMetaData(result: any): string {
      </html>
    `;
 }
-
-export function getClusterConnectingFormView(): string {
-   return `
-      <!DOCTYPE html>
-      <html>
-      <head>
-         <meta charset="UTF-8">
-         <title>Add Connection</title>
-      </head>
-      <body>
-         <form>
-            <label for="url">Enter Cluter Connection URL:</label><br>
-            <input type="text" id="url" name="url" placeholder="URL" value="couchbase://localhost" required /> <br/><br/>
-            <label for="url">Enter Username:</label><br>
-            <input type="text" id="username" name="class" placeholder="Username" value="Administrator" required /> <br/><br/>
-            <label for="password">Enter Password:</label><br> 
-            <input type="password" id="password" name="rollno" placeholder="Password" value="password" required> <br/><br/>
-            <label for="connectionIndetifier">Enter Connection Identifier (optional):</label><br>
-            <input type="text" id="connectionIdentifier" name="identifier" placeHolder="Connection Identifier" value="" > <br/><br/>
-            <input type="button" value="Submit" onclick="postRequestToConnect()"/>
-         </form>
-         <script>
-         function postRequestToConnect(){
-            const vscode = acquireVsCodeApi();
-            let url = document.getElementById('url').value;
-            let username = document.getElementById('username').value;
-            let password = document.getElementById('password').value;
-            let identifier = document.getElementById('connectionIdentifier').value;
-            vscode.postMessage({
-            command: 'submit',
-            url: url,
-            username: username,
-            password: password,
-            connectionIdentifier: identifier
-         })};
-      </script>
-      </body>
-      </html>
-    `;
-}
